+++ PULL_REQUEST_TEMPLATE.md
@@ -10,7 +10,7 @@ Fixes # (issue)
 Please delete options that are not relevant.
 
 - [ ] New Script
-- [ ] New Category (Is any new category being made to accomodate this script)
+- [ ] New Category (Is any new category being made to accommodate this script)
 - [ ] Bug fix (non-breaking change which fixes an issue)
 - [ ] New feature (non-breaking change which adds functionality)
 ```
